Move the Request a Site form under "/Forms" module
